refactor(/package/qa): move components about draft from /libs/draft-js to /react-components/draft
