+++ packages/qa/src/react-components/draft/color-text.js
