+++ packages/qa/src/react-components/draft/entity-decorator.js
