+++ packages/qa/src/react-components/draft/image.js
